+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for Eurocircuits drill adjust to fix small OARs.
+  Option `eurocircuits_reduce_holes`. (#227)
 
 
 ## [1.3.0] - 2022-09-08